1.增加每日运营-》用户预览 2.优化每日运营-》任务执行代码

+++ operate/src/main/java/com/linksteady/operate/controller/DailyController.java
@@ -1,9 +1,12 @@
 package com.linksteady.operate.controller;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
+import com.alibaba.fastjson.JSONArray;
+import com.alibaba.fastjson.JSONObject;
 import com.linksteady.common.domain.QueryRequest;
 import com.linksteady.common.domain.ResponseBo;
 import com.linksteady.common.util.FileUtils;
+import com.linksteady.operate.config.ConfigCacheManager;
 import com.linksteady.operate.domain.*;
 import com.linksteady.operate.service.*;
 import com.linksteady.operate.vo.DailyPersonalVo;
@@ -11,13 +14,13 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
+
+import java.math.BigDecimal;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.*;
-import java.util.concurrent.locks.LockSupport;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
@@ -54,7 +57,15 @@ public class DailyController {
         String touchDt = request.getParam().get("touchDt");
         int count = dailyService.getTotalCount(touchDt);
         List<DailyHead> dailyInfos = dailyService.getPageList(start, end, touchDt);
-        return ResponseBo.okOverPaging(null, count, dailyInfos);
+
+        //设置当前天记录的 校验状态
+        String validateLabel=dailyService.validUserGroup()?"未通过":"通过";
+
+        dailyInfos.stream().forEach(p->{
+           p.setValidateLabel(validateLabel);
+        });
+
+        return ResponseBo.okOverPaging(validateLabel, count, dailyInfos);
     }
 
     /**
@@ -75,16 +86,6 @@ public class DailyController {
         return ResponseBo.okOverPaging(null, count, dataList);
     }
 
-    /**
-     * 获取编辑页xxx，共x人，选择y人
-     *
-     * @param headId
-     * @return
-     */
-    @GetMapping("/getTipInfo")
-    public ResponseBo getTipInfo(@RequestParam String headId) {
-        return ResponseBo.okWithData(null, dailyService.getTipInfo(headId));
-    }
 
     /**
      * 生成待推送的短信文案
@@ -93,26 +94,34 @@ public class DailyController {
      */
     @GetMapping("/generatePushList")
     public ResponseBo generatePushList(String headId) {
-        String result;
-        try {
+
+        //首先获取锁
+        if(dailyService.getTransContentLock(headId))
+        {
+            //调用文案生成逻辑
             dailyDetailService.deletePushContentTemp(headId);
             //1表示生成成功 0表示生成失败
-            result = dailyDetailService.generatePushList(headId);
-
+            String result = dailyDetailService.generatePushList(headId);
             if ("1".equals(result)) {
-                return ResponseBo.ok();
+                //释放锁
+                dailyService.delTransLock();
+
+                long timestamp=System.currentTimeMillis();
+                //更新时间戳到头表
+                dailyService.updateHeaderOpChangeDate(headId,timestamp);
+
+                return ResponseBo.ok(timestamp);
             } else {
                 return ResponseBo.error("生成文案错误，请联系系统运维人员！");
             }
-
-        } catch (Exception e) {
-            e.printStackTrace();
-            return ResponseBo.error("策略生成错误，请检查配置！");
+        }else
+        {
+            return ResponseBo.error("其他用户正在生成文案，请稍后再操作！");
         }
     }
 
     /**
-     * 获取用户成长策略表（第二步）
+     * 获取用户的文案
      *
      * @return
      */
@@ -127,14 +136,177 @@ public class DailyController {
     }
 
     /**
+     * 获取用户预览数据
+     *
+     * @return
+     */
+    @GetMapping("/getUserStatsData")
+    public ResponseBo getUserStatsData(String headId) {
+        Map<String,String> pathActiveMap= ConfigCacheManager.getInstance().getPathActiveMap();
+        Map<String,String> userValueMap=ConfigCacheManager.getInstance().getUserValueMap();
+        Map<String,String> lifeCycleMap=ConfigCacheManager.getInstance().getLifeCycleMap();
+
+        //获取人群分布
+        List<DailyUserStats> dailyUserStats=dailyService.getUserStats(headId);
+
+       dailyUserStats.stream().forEach(p->{
+           p.setUserValueLabel(userValueMap.get(p.getUserValue()));
+           p.setGetPathActivityLabel(pathActiveMap.get(p.getPathActivity()));
+           p.setLifecycleLabel(lifeCycleMap.get(p.getLifecycle()));
+       });
+
+        JSONArray fparray=new JSONArray();
+        JSONArray rparray=new JSONArray();
+        JSONArray fpTemp=null;
+        for(DailyUserStats d1:dailyUserStats)
+        {
+            fpTemp=new JSONArray();
+            /**
+             * x轴的数据 人数
+             */
+            fpTemp.add(d1.getUcnt());
+            /**
+             * y轴 用户活跃度
+             */
+            fpTemp.add(d1.getGetPathActivityLabel());
+
+            int formatSize=0;
+            if("ULC_01".equals(d1.getUserValue()))
+            {
+                formatSize=80;
+            }else if("ULC_02".equals(d1.getUserValue()))
+            {
+                formatSize=60;
+            }else if("ULC_03".equals(d1.getUserValue()))
+            {
+                formatSize=40;
+            }else if("ULC_04".equals(d1.getUserValue()))
+            {
+                formatSize=20;
+            }
+            fpTemp.add(formatSize);
+            fpTemp.add(d1.getLifecycleLabel());
+            fpTemp.add("人数:"+d1.getUcnt()+",占比:"+d1.getPct()+"%");
+            fpTemp.add(d1.getUserValue());
+            fpTemp.add(d1.getPathActivity());
+            fpTemp.add(d1.getLifecycle());
+
+            if("1".equals(d1.getLifecycle()))
+            {
+                fparray.add(fpTemp);
+            }else
+            {
+                rparray.add(fpTemp);
+            }
+        }
+
+        Map<String,Object> result=Maps.newHashMap();
+        result.put("fpUser",fparray);
+        result.put("rpUser",rparray);
+
+        //获取排名第一的群组 获取其类目信息
+        DailyUserStats dailyUserStats1=dailyUserStats.get(0);
+        if(null!=dailyUserStats1)
+        {
+            //获取用户在SPU上top 10
+            List<DailyUserStats> spuList=dailyService.getUserStatsBySpu(headId,dailyUserStats1.getUserValue(),dailyUserStats1.getPathActivity(),dailyUserStats1.getLifecycle());
+            result.put("spuList",spuList);
+            result.put("groupName",dailyUserStats1.getUserValueLabel()+","+dailyUserStats1.getGetPathActivityLabel()+","+dailyUserStats1.getLifecycleLabel()+"群组");
+            result.put("userValue",dailyUserStats1.getUserValue());
+            result.put("pathActive",dailyUserStats1.getPathActivity());
+            result.put("lifecycle",dailyUserStats1.getLifecycle());
+
+            DailyUserStats dailyUserStats2=spuList.get(0);
+
+            //获取top10 推荐商品
+            if(null!=dailyUserStats2)
+            {
+                List<DailyUserStats> prodList=dailyService.getUserStatsByProd(headId,dailyUserStats1.getUserValue(),dailyUserStats1.getPathActivity(),dailyUserStats1.getLifecycle(),dailyUserStats2.getSpuName());
+                result.put("prodList",prodList);
+                result.put("prodGroupName",dailyUserStats1.getUserValueLabel()+","+dailyUserStats1.getGetPathActivityLabel()+","+dailyUserStats1.getLifecycleLabel()+"群组"+dailyUserStats2.getSpuName()+"类目");
+                result.put("spuName",dailyUserStats2.getSpuName());
+            }else
+            {
+                result.put("prodList",Lists.newArrayList());
+                result.put("prodGroupName","");
+                result.put("spuName","");
+            }
+
+        }else
+        {
+            result.put("spuList",Lists.newArrayList());
+            result.put("groupName","");
+            result.put("userValue","");
+            result.put("pathActive","");
+            result.put("lifecycle","");
+        }
+        return ResponseBo.okWithData("",result);
+    }
+
+    /**
+     * 用户预览 刷新SPU表格和PROD表格
+     */
+    @GetMapping("/refreshUserStatData")
+    public ResponseBo refreshUserStatData(String headId,String userValue,String pathActive,String lifecycle) {
+        Map<String,String> pathActiveMap= ConfigCacheManager.getInstance().getPathActiveMap();
+        Map<String,String> userValueMap=ConfigCacheManager.getInstance().getUserValueMap();
+        Map<String,String> lifeCycleMap=ConfigCacheManager.getInstance().getLifeCycleMap();
+
+        Map<String,Object> result=Maps.newHashMap();
+
+        List<DailyUserStats> spuList=dailyService.getUserStatsBySpu(headId,userValue,pathActive,lifecycle);
+        result.put("spuList",spuList);
+        result.put("groupName",userValueMap.get(userValue)+","+pathActiveMap.get(pathActive)+","+lifeCycleMap.get(lifecycle)+"群组");
+
+        DailyUserStats dailyUserStats2=spuList.get(0);
+
+        //获取top10 推荐商品
+        if(null!=dailyUserStats2)
+        {
+            List<DailyUserStats> prodList=dailyService.getUserStatsByProd(headId,userValue,pathActive,lifecycle,dailyUserStats2.getSpuName());
+            result.put("prodList",prodList);
+            result.put("prodGroupName",userValueMap.get(userValue)+","+pathActiveMap.get(pathActive)+","+lifeCycleMap.get(lifecycle)+"群组"+dailyUserStats2.getSpuName()+"类目");
+            result.put("spuName",dailyUserStats2.getSpuName());
+        }else
+        {
+            result.put("prodList",Lists.newArrayList());
+            result.put("prodGroupName","");
+            result.put("spuName","");
+        }
+        return ResponseBo.okWithData("",result);
+    }
+
+    /**
+     * 用户预览 刷新PROD表格
+     */
+    @GetMapping("/refreshUserStatData2")
+    public ResponseBo refreshUserStatData2(String headId,String userValue,String pathActive,String lifecycle,String spuName) {
+
+        Map<String,String> pathActiveMap= ConfigCacheManager.getInstance().getPathActiveMap();
+        Map<String,String> userValueMap=ConfigCacheManager.getInstance().getUserValueMap();
+        Map<String,String> lifeCycleMap=ConfigCacheManager.getInstance().getLifeCycleMap();
+
+        Map<String,Object> result=Maps.newHashMap();
+
+        List<DailyUserStats> prodList=dailyService.getUserStatsByProd(headId,userValue,pathActive,lifecycle,spuName);
+        result.put("prodList",prodList);
+        result.put("prodGroupName",userValueMap.get(userValue)+","+pathActiveMap.get(pathActive)+","+lifeCycleMap.get(lifecycle)+"群组"+spuName+"类目");
+
+        return ResponseBo.okWithData("",result);
+    }
+
+
+
+    /**
      * 启动群组推送
      *
      * @return
      */
     @GetMapping("/submitData")
     @Transactional(rollbackFor = Exception.class)
-    public synchronized ResponseBo submitData(String headId, String pushMethod, String pushPeriod) {
-        String status = dailyService.getStatusById(headId);
+    public synchronized ResponseBo submitData(String headId, String pushMethod, String pushPeriod,Long timestamp) {
+        //进行一次状态的判断
+        String status = dailyService.getDailyHeadById(headId).getStatus();
         if (!status.equalsIgnoreCase("todo")) {
             return ResponseBo.error("当前数据状态不支持该操作！");
         }
@@ -144,16 +316,29 @@ public class DailyController {
         if (null != validResult) {
             return ResponseBo.error(validResult);
         }
-        // 推送方式 IMME立即推送 AI智能推送 FIXED固定时间推送
-        updateSmsPushMethod(headId, pushMethod, pushPeriod);
 
-        //复制写入待推送列表
-        dailyDetailService.copyToPushList(headId);
+        //判断时间戳
+        int count=dailyService.validateOpChangeTime(headId,timestamp);
+        if(count==0)
+        {
+            return ResponseBo.error("数据已被其他用户操作，请返回列表界面重新尝试！");
+        }else
+        {
+            //更新时间戳
+            long timestampNew=System.currentTimeMillis();
+            //更新时间戳到头表
+            dailyService.updateHeaderOpChangeDate(headId,timestampNew);
 
-        status = "done";
-        dailyService.updateStatus(headId, status);
-        return ResponseBo.ok();
+            // 推送方式 IMME立即推送 AI智能推送 FIXED固定时间推送
+            updateSmsPushMethod(headId, pushMethod, pushPeriod);
 
+            //复制写入待推送列表
+            dailyDetailService.copyToPushList(headId);
+
+            status = "done";
+            dailyService.updateStatus(headId, status);
+            return ResponseBo.ok();
+        }
     }
 
     /**
@@ -206,7 +391,7 @@ public class DailyController {
      */
     @GetMapping("/getStatusById")
     public ResponseBo getStatusById(@RequestParam String headId) {
-        return ResponseBo.okWithData(null, dailyService.getStatusById(headId));
+        return ResponseBo.okWithData(null, dailyService.getDailyHeadById(headId).getStatus());
     }
 
     /**
@@ -232,7 +417,7 @@ public class DailyController {
     /**
      * 获取用户组配置分页数据
      *
-     * @param request
+     * @param
      * @return
      */
     @GetMapping("/userGroupList")
@@ -338,7 +523,7 @@ public class DailyController {
     }
 
     /**
-     * 触达用户之前进行用户群组的验证
+     * 触达用户之前进行用户群组的验证 并返回验证结果 true表示验证未通过 false表示验证通过
      *
      * @return
      */
